+++ src/main/Mock/Bitrix/Main/DB/SqlExpression.php
@@ -1,9 +1,38 @@
 <?php
+/** @noinspection PhpDocRedundantThrowsInspection */
 
 namespace WebArch\BitrixTaxidermist\Mock\Bitrix\Main\DB;
 
+use WebArch\BitrixTaxidermist\Mock\Bitrix\Main\ArgumentException;
+
 class SqlExpression
 {
+    /** @var string */
+    protected $expression;
+
+    /** @var array */
+    protected $args = [];
+
+    protected $pattern = '/([^\\\\]|^)(\?[#sif]?)/';
+
+    protected $i;
+
+    /**
+     * @var Connection
+     */
+    private $connection;
+
+    /**
+     * @param string $expression Sql expression.
+     * @param string,... $args Substitutes.
+     *
+     * @throws ArgumentException
+     * @noinspection PhpDocSignatureInspection
+     */
+    public function __construct(string $expression, string ...$args)
+    {
+    }
+
     /**
      * Returns $expression with replaced placeholders.
      *
@@ -12,5 +41,41 @@ class SqlExpression
     public function compile()
     {
         return '';
+    }
+
+    /**
+     * Used by compile method to replace placeholders with values.
+     *
+     * @param array $matches Matches found by preg_replace.
+     *
+     * @return string
+     * @noinspection PhpMissingParamTypeInspection
+     * @noinspection PhpUnusedParameterInspection
+     */
+    protected function execPlaceholders($matches)
+    {
+        return '';
+    }
+
+    public function __toString()
+    {
+        return $this->compile();
+    }
+
+    /**
+     * @return Connection
+     */
+    public function getConnection()
+    {
+        return $this->connection;
+    }
+
+    /**
+     * @param Connection $connection
+     *
+     * @noinspection PhpMissingParamTypeInspection*/
+    public function setConnection($connection)
+    {
+        $this->connection = $connection;
     }
 }