+++ src/main/Mock/Bitrix/Main/ORM/Fields/Field.php
@@ -14,11 +14,15 @@ abstract class Field
     /** @var Entity */
     protected $entity;
 
+    /** @var null|callback[] */
+    protected $saveDataModifiers;
+
     /**
      * @param string $name
      * @param array<string, mixed> $parameters deprecated, use configure* and add* methods instead
      *
      * @throws SystemException
+     * @noinspection PhpMissingParamTypeInspection
      */
     public function __construct($name, $parameters = [])
     {
@@ -30,6 +34,7 @@ abstract class Field
      * @throws SystemException
      * @return Field
      * @noinspection PhpUnusedParameterInspection
+     * @noinspection PhpMissingParamTypeInspection
      */
     public function addFetchDataModifier($modifier)
     {
@@ -59,5 +64,27 @@ abstract class Field
     public function getEntity()
     {
         return $this->entity;
+    }
+
+    /**
+     * @throws SystemException
+     * @return null|array|callback[]
+     */
+    public function getSaveDataModifiers()
+    {
+        return $this->saveDataModifiers;
+    }
+
+    /**
+     * @param $modifier
+     *
+     * @throws SystemException
+     * @return $this
+     * @noinspection PhpUnusedParameterInspection
+     * @noinspection PhpUnusedParameterInspection
+     */
+    public function addSaveDataModifier($modifier)
+    {
+        return $this;
     }
 }